requirements.txt
requests>=2.27.1
beautifulsoup4>=4.11.1
dnspython>=2.2.1
python-whois>=0.7.3
nmap>=0.6.1
colorama>=0.4.5
pystyle>=2.9
scapy>=2.4.5
PyPDF2>=2.10.5
Pillow>=9.2.0
python-docx>=0.8.11
mutagen>=1.45.1
ipaddress>=1.0.23
urllib3>=1.26.9
python-nmap>=0.7.1
cryptography>=39.0.2
pyOpenSSL>=23.0.0
tqdm>=4.64.0
python-dateutil>=2.8.2
validators>=0.20.0
fake-useragent>=1.1.3
lxml>=4.9.2
pymongo>=4.3.3
psutil>=5.9.4
pytz>=2022.7.1
shodan>=1.28.0
tldextract>=3.4.0
selenium>=4.8.2
webdriver-manager>=3.8.5